New PR from HU to Main

+++ Application/Services/CustomerService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -9,9 +10,11 @@ using Application.Common.FluentValidations.Extentions;
 using Application.Common.FluentValidations.Validators;
 using Application.Common.Helpers.Exceptions;
 using Application.DTOs;
+using Application.DTOs.Entries;
 using Application.Interfaces.Infrastructure.Mongo;
 using Application.Interfaces.Services;
 using Common.Helpers.Exceptions;
+using Core.Entities.MongoDB;
 using Microsoft.Extensions.Logging;
 
 namespace Application.Services
@@ -32,29 +35,57 @@ namespace Application.Services
             _customerRepository = customerRepository;
             _logger = logger;
         }
+        /// <summary>
+        /// Get the document types
+        /// </summary>
+        /// <returns></returns>
+        public Dictionary<string, string> GetDocumentTypes()
+        {
+            Dictionary<string, string> keyValues = new Dictionary<string, string>()
+            {
+                {"TI", ""},
+                {"CC", ""},
+                {"NIT", ""}
+            };
+            return keyValues;
+        }
 
         /// <summary>
         /// Private method controls the process of create a product
         /// </summary>
-        /// <param name="customer"></param>
+        /// <param name="customerInput"></param>
         /// <returns></returns>
         /// <exception cref="BusinessException"></exception>
-        public async Task CreateCustomer(Customer customer)
+        public async Task<CustomerCollection> CreateCustomer(CustomerInput customerInput)
         {
             try
             {
-                await customer.ValidateAndThrowsAsync<Customer, CustomerValidator>();
-                await _customerRepository.CreateCustomerAsync(customer);
+                await customerInput.ValidateAndThrowsAsync<CustomerInput, CustomerValidator>();
+                Dictionary<string, string> keyValues = GetDocumentTypes();
+                CustomerOutput customer = new()
+                {
+                    DocumentType = customerInput.DocumentType,
+                    Document = customerInput.Document,
+                    Name = customerInput.Name,
+                    Email = customerInput.Email,
+                    Phone = customerInput.Phone
+                };
+                if (keyValues.ContainsKey(customerInput.DocumentType))
+                    return await _customerRepository.CreateAsync(customer);
+                else
+                    throw new BusinessException(nameof(GateWayBusinessException.CustomerDocumentTypeIsInvalid),
+                    nameof(GateWayBusinessException.CustomerDocumentTypeIsInvalid));
+
             }
             catch (BusinessException bex)
             {
-                _logger.LogError(bex, "Error: {message} Error Code: {code-message} creating customer: {customer}"
-                    , bex.Code, bex.Message, customer);
+                _logger.LogError(bex, "Error: {message} Error Code: {code-message}"
+                    , bex.Code, bex.Message);
                 throw new BusinessException(bex.Message, bex.Code);
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error: {message} creating customer: {customer} ", ex.Message, customer);
+                _logger.LogError(ex, "Error: {message}", ex.Message);
                 throw new BusinessException(nameof(GateWayBusinessException.NotControlledException),
                     nameof(GateWayBusinessException.NotControlledException));
             }
@@ -65,12 +96,15 @@ namespace Application.Services
         /// <param name="_id"></param>
         /// <returns></returns>
         /// <exception cref="BusinessException"></exception>
-        public async Task<Customer> GetCustomerById(string _id)
+        public async Task<CustomerOutput> GetCustomerById(string _id)
         {
             try
             {
-                var result = await _customerRepository.GetCustomerByIdAsync(_id);
-                return result;
+                if (!String.IsNullOrEmpty(_id))                    
+                    return await _customerRepository.GetCustomerByIdAsync(_id);                
+                else
+                    throw new BusinessException(nameof(GateWayBusinessException.CustomerIdCannotBeNull),
+                    nameof(GateWayBusinessException.CustomerIdCannotBeNull));
             }
             catch (BusinessException bex)
             {
@@ -89,18 +123,36 @@ namespace Application.Services
         /// <summary>
         /// Private method controls the process of create a product
         /// </summary>
-        /// <param name="customer"></param>
+        /// <param name="customerInput"></param>
+        /// <param name="_id"></param>
         /// <returns></returns>
         /// <exception cref="BusinessException"></exception>
-        public async Task UpdateCustomer(Customer customer)
+        public async Task<CustomerOutput> UpdateCustomerData(CustomerInput customerInput, string _id)
         {
             try
             {
-                await customer.ValidateAndThrowsAsync<Customer, CustomerValidator>();
-                var update = await _customerRepository.UpdateCustomerAsync(customer);
-                if (update == false)
-                    throw new BusinessException(nameof(GateWayBusinessException.CustomerIdIsNotValid),
-                    nameof(GateWayBusinessException.CustomerIdIsNotValid));
+                await customerInput.ValidateAndThrowsAsync<CustomerInput, CustomerValidator>();
+                Dictionary<string, string> keyValues = GetDocumentTypes();
+                CustomerOutput customer = new()
+                {
+                    _id = _id,
+                    DocumentType = customerInput.DocumentType,
+                    Document = customerInput.Document,
+                    Name = customerInput.Name,
+                    Email = customerInput.Email,
+                    Phone = customerInput.Phone
+                };
+                if (!String.IsNullOrEmpty(_id))
+                {
+                    if (keyValues.ContainsKey(customerInput.DocumentType))
+                        return await _customerRepository.UpdateCustomerDataAsync(customer);
+                    else
+                        throw new BusinessException(nameof(GateWayBusinessException.CustomerDocumentTypeIsInvalid),
+                        nameof(GateWayBusinessException.CustomerDocumentTypeIsInvalid));
+                }
+                else
+                    throw new BusinessException(nameof(GateWayBusinessException.CustomerIdCannotBeNull),
+                    nameof(GateWayBusinessException.CustomerIdCannotBeNull));
             }
             catch (BusinessException bex)
             {
@@ -126,12 +178,19 @@ namespace Application.Services
         {
             try
             {
-                var delete = await _customerRepository.DeleteCustomerAsync(_id);
-                if (delete == false)
-                    throw new BusinessException(nameof(GateWayBusinessException.CustomerIdIsNotValid),
-                    nameof(GateWayBusinessException.CustomerIdIsNotValid));
+                if (!String.IsNullOrEmpty(_id))
+                    await _customerRepository.DeleteCustomerAsync(_id);
+                else
+                    throw new BusinessException(nameof(GateWayBusinessException.CustomerIdCannotBeNull),
+                    nameof(GateWayBusinessException.CustomerIdCannotBeNull));
             }
-            catch (FormatException ex)
+            catch (BusinessException bex)
+            {
+                _logger.LogError(bex, "Error: {message} Error Code: {code-message}"
+                    , bex.Code, bex.Message);
+                throw new BusinessException(bex.Message, bex.Code);
+            }
+            catch (Exception ex)
             {
                 _logger.LogError(ex, "Error: {message}", ex.Message);
                 throw new BusinessException(nameof(GateWayBusinessException.CustomerIdIsNotValid),
